requirements.txt
requests
spacy
python-dotenv
dateparser
SpeechRecognition
PyAudio
pyttsx3
PySide6
